Allow changing warpFactorLevels (in warp.m) using the ARTwarp Menu
It is tested on warpfactor levels 3,4,5,100 and it appears to work without crashing. In general,
as the warp factor increases (which kind of means it allows more streaching and squishing along the
time axis) the number of categories go down. This is might be a flawed observation, but it makes sense
because warping whistles on a larger scale (larger warp level) will make them look similar, leading
to fewer groups being made.

Problems:

    - When the level is set to 0,1, or 2, it throws an error saying it is unable to unwarp a
    whistle.

    - There is time warping and frequency warping taking place on both axis according to the
    comments at the start of the 'warp.m' script. So it is a bit ambiguous which 3 refers
    to which axis in the code. I have just replaced most of the 3s with the variable
    'warpFactorLevel'.

Attempts to resolve issue #12