Initial CI setup
This adds two GitHub Actions workflows:
- for script tests
- for unit tests

and two example tests, both of them are just opening a CSV file,
just to check that the wokflows work (or not, as I have checked
in my fork). This is the foundation that allows us to enable CI
for the ARTwarp repository and build further system of tests and
workflows.